1. add `processCharacter` module 2. add `setCycle` and `setTurned` actionCreators 3. init game state 4. subscribe stateListeners 5. register `characterState`
