+++ src/index.js
@@ -2,155 +2,31 @@ import './styles/main.scss'
 
 import store from 'redux/store'
 
-import characterAtlas from './assets/Male-3-Walk.png' // 3 x 4
 // import terrainAtlas from './assets/terrain.png'
 
-import processControls from './modules/processControls'
+import processControls from 'modules/processControls'
+import processCharacter from 'modules/processCharacter'
 
 const canvas = document.querySelector('#game')
 const { width: w, height: h } = canvas.getBoundingClientRect()
 const ctx = canvas.getContext('2d')
 
-const img = document.createElement('img')
-img.src = characterAtlas
-
-let keyState = {}
-
-const handleImgLoad = () => {
-  const sprite = {
-    w: 48,
-    h: 48,
-    shots: 3,
-    step: 10,
-  }
-
-  let spriteState = {
-    cycle: 0,
-    posX: w / 2 - sprite.w / 2,
-    posY: h / 2 - sprite.h / 2,
-    turned: 'down',
-  }
-
-  const directionsMap = {
-    down: 0,
-    left: 1,
-    right: 2,
-    up: 3,
-  }
-
-  const animateSprite = () => {
-    spriteState = {
-      ...spriteState,
-      cycle: (spriteState.cycle + 1) % sprite.shots,
-    }
-  }
-
-  const moveTo = (direction) => {
-    const posMap = {
-      left: 'posX',
-      right: 'posX',
-      up: 'posY',
-      down: 'posY',
-    }
-
-    const moveToMap = {
-      left: () => spriteState.posX - (spriteState.posX <= 0 ? 0 : sprite.step),
-      right: () => spriteState.posX + (spriteState.posX + sprite.w >= w ? 0 : sprite.step),
-      up: () => spriteState.posY - (spriteState.posY <= 0 ? 0 : sprite.step),
-      down: () => spriteState.posY + (spriteState.posY + sprite.h >= h ? 0 : sprite.step),
-    }
-
-    spriteState = {
-      ...spriteState,
-      [posMap[direction]]: moveToMap[direction](),
-    }
-  }
-
-  const turnTo = (direction) => {
-    spriteState = {
-      ...spriteState,
-      turned: direction,
-    }
-  }
-
-  setInterval(() => {
-    switch (true) {
-      case keyState.isLeftPressed && keyState.isUpPressed: {
-        turnTo('left')
-        moveTo('left')
-        moveTo('up')
-        animateSprite()
-        break
-      }
-      case keyState.isRightPressed && keyState.isUpPressed: {
-        turnTo('right')
-        moveTo('right')
-        moveTo('up')
-        animateSprite()
-        break
-      }
-      case keyState.isRightPressed && keyState.isDownPressed: {
-        turnTo('right')
-        moveTo('right')
-        moveTo('down')
-        animateSprite()
-        break
-      }
-      case keyState.isLeftPressed && keyState.isDownPressed: {
-        turnTo('left')
-        moveTo('left')
-        moveTo('down')
-        animateSprite()
-        break
-      }
-      case keyState.isRightPressed: {
-        turnTo('right')
-        moveTo('right')
-        animateSprite()
-        break
-      }
-      case keyState.isLeftPressed: {
-        turnTo('left')
-        moveTo('left')
-        animateSprite()
-        break
-      }
-      case keyState.isUpPressed: {
-        turnTo('up')
-        moveTo('up')
-        animateSprite()
-        break
-      }
-      case keyState.isDownPressed: {
-        turnTo('down')
-        moveTo('down')
-        animateSprite()
-        break
-      }
-      default:
-        break
-    }
-
-    ctx.clearRect(0, 0, w, h)
-
-    ctx.drawImage(
-      img, // img elem
-      spriteState.cycle * sprite.w, // image pos x
-      sprite.h * directionsMap[spriteState.turned], // image pos y
-      sprite.w, // image width
-      sprite.h, // image height
-      spriteState.posX, // canvas pos x
-      spriteState.posY, // canvas pos y
-      sprite.w, // canvas image width
-      sprite.h, // canvas image height
-    )
-  }, 100)
+const state = {
+  key: store.getState().keyState,
+  character: store.getState().characterState,
 }
 
-img.addEventListener('load', handleImgLoad)
+const keyStateListenter = () => {
+  state.key = store.getState().keyState
+}
+const characterStateListenter = () => {
+  state.character = store.getState().characterState
+}
 
-store.subscribe(() => {
-  keyState = store.getState().keyState
-})
+store.subscribe(keyStateListenter)
+store.subscribe(characterStateListenter)
 
 processControls()
+processCharacter({
+  w, h, ctx, state,
+})
