+++ src/redux/store.js
@@ -1,8 +1,10 @@
 import { createStore, combineReducers } from 'redux'
 import keyState from './reducers/keyState'
+import characterState from './reducers/characterState'
 
 const rootReducer = combineReducers({
   keyState,
+  characterState,
 })
 
 const store = createStore(rootReducer)
